Split up large controller, remove repeated logic
