+++ src/app/src/main/kotlin/org/imperial/mrc/hint/AppProperties.kt
@@ -43,6 +43,7 @@ interface AppProperties
     val oauth2ClientAudience: String
     val oauth2ClientScope: String
     val oauth2LoginMethod: Boolean
+    val fileTypeMappings: Map<String, String>
 }
 
 //prevent auto-wiring of default Properties
@@ -51,41 +52,53 @@ class HintProperties : Properties()
 @Component
 class ConfiguredAppProperties(private val props: HintProperties = properties) : AppProperties
 {
-    override val adrANCSchema = propString("adr_anc_schema")
-    override val adrARTSchema = propString("adr_art_schema")
-    override val adrPJNZSchema = propString("adr_pjnz_schema")
-    override val adrPopSchema = propString("adr_pop_schema")
-    override val adrShapeSchema = propString("adr_shape_schema")
-    override val adrSurveySchema = propString("adr_survey_schema")
-    override val adrVmmcSchema = propString("adr_vmmc_schema")
-    override val adrOutputZipSchema = propString("adr_output_zip_schema")
-    override val adrOutputSummarySchema = propString("adr_output_summary_schema")
-    override val adrOutputComparisonSchema = propString("adr_output_comparison_schema")
-    override val adrDatasetSchema = propString("adr_schema")
-    override val adrUrl = propString("adr_url")
-    override val apiUrl = propString("hintr_url")
-    override val applicationTitle = propString("application_title")
-    override val applicationUrl = propString("application_url")
-    override val emailMode = propString("email_mode")
-    override val emailServer = propString("email_server")
-    override val emailPort = propString("email_port").toIntOrNull()
-    override val emailSender = propString("email_sender")
-    override val emailUsername = propString("email_username")
-    override val emailPassword = propString("email_password")
-    override val tokenIssuer = propString("token_issuer")
-    override val uploadDirectory = propString("upload_dir")
-    override val dbUser: String = propString("db_user")
-    override val dbPassword: String = propString("db_password")
-    override val dbUrl: String = propString("db_url")
-    override val issueReportUrl: String = propString("issue_report_url")
-    override val supportEmail: String = propString("support_email")
-    override val oauth2ClientId: String = propString("oauth2_client_id")
-    override val oauth2ClientSecret: String = propString("oauth2_client_secret")
-    override val oauth2ClientUrl: String = propString("oauth2_client_url")
-    override val oauth2ClientAudience: String = propString("oauth2_client_audience")
-    override val oauth2ClientAdrServerUrl = propString("oauth2_client_adr_server_url")
-    override val oauth2ClientScope = propString("oauth2_client_scope")
-    override val oauth2LoginMethod = propString("oauth2_login_method").toBoolean()
+    final override val adrANCSchema = propString("adr_anc_schema")
+    final override val adrARTSchema = propString("adr_art_schema")
+    final override val adrPJNZSchema = propString("adr_pjnz_schema")
+    final override val adrPopSchema = propString("adr_pop_schema")
+    final override val adrShapeSchema = propString("adr_shape_schema")
+    final override val adrSurveySchema = propString("adr_survey_schema")
+    final override val adrVmmcSchema = propString("adr_vmmc_schema")
+    final override val adrOutputZipSchema = propString("adr_output_zip_schema")
+    final override val adrOutputSummarySchema = propString("adr_output_summary_schema")
+    final override val adrOutputComparisonSchema = propString("adr_output_comparison_schema")
+    final override val adrDatasetSchema = propString("adr_schema")
+    final override val adrUrl = propString("adr_url")
+    final override val apiUrl = propString("hintr_url")
+    final override val applicationTitle = propString("application_title")
+    final override val applicationUrl = propString("application_url")
+    final override val emailMode = propString("email_mode")
+    final override val emailServer = propString("email_server")
+    final override val emailPort = propString("email_port").toIntOrNull()
+    final override val emailSender = propString("email_sender")
+    final override val emailUsername = propString("email_username")
+    final override val emailPassword = propString("email_password")
+    final override val tokenIssuer = propString("token_issuer")
+    final override val uploadDirectory = propString("upload_dir")
+    final override val dbUser: String = propString("db_user")
+    final override val dbPassword: String = propString("db_password")
+    final override val dbUrl: String = propString("db_url")
+    final override val issueReportUrl: String = propString("issue_report_url")
+    final override val supportEmail: String = propString("support_email")
+    final override val oauth2ClientId: String = propString("oauth2_client_id")
+    final override val oauth2ClientSecret: String = propString("oauth2_client_secret")
+    final override val oauth2ClientUrl: String = propString("oauth2_client_url")
+    final override val oauth2ClientAudience: String = propString("oauth2_client_audience")
+    final override val oauth2ClientAdrServerUrl = propString("oauth2_client_adr_server_url")
+    final override val oauth2ClientScope = propString("oauth2_client_scope")
+    final override val oauth2LoginMethod = propString("oauth2_login_method").toBoolean()
+    final override val fileTypeMappings = mapOf(
+            "baseUrl" to adrUrl,
+            "anc" to adrANCSchema,
+            "programme" to adrARTSchema,
+            "pjnz" to adrPJNZSchema,
+            "population" to adrPopSchema,
+            "shape" to adrShapeSchema,
+            "survey" to adrSurveySchema,
+            "vmmc" to adrVmmcSchema,
+            "outputZip" to adrOutputZipSchema,
+            "outputSummary" to adrOutputSummarySchema,
+            "outputComparison" to adrOutputComparisonSchema)
 
     companion object
     {
